Fixed a few bugs that showed only in exports
Worked very well when loauching through Eclipse, but when exported the product, wouldn't work.
It happened to be due to Java versions issues (needed fields to be declared final, in some event/jobs contexts
but wouldn't report an error...). Switching the project to Java 1.8+ did solve it

So now on, the project meeds Java 1.8+
======================================